fix: Fix bugs in actions and AsyncStorage and connect View
Fixed a bug in actions, where there was either a wrong return type (an action needs to return an object) or no return at all.
Fixed a bug in AsyncStorage, because decks were not being stored in the DB.
Replaced the `submitEntry` function in the `NewDeck` View with `submitNewDeck`. Changed the parameter type in that function to take an object and a key (previously it was getting undefined as object) and refactored the function inside the `api` file to merge the new item and log the result to the console. Added a call to AsyncStorage to log all the stored keys and values, too.
Hooked up Redux to the application, connecting the View to dispatch an action as a user clicked the `submit` button.