feat: Add functionality to New Question View
Added two input fields and a submit button to the New Question View. The submit button just logs the text to the console at the moment.
An alert message has been added, which only displays if the user tries to submit one of the fields (or both) without any content in it.