feat: Implement quiz mechanism
Coded quiz mechanism, letting the app display the questions from the cards inserted in the selected deck. At the moment it is only displayed the question text, without any styles added.
Made the `Start Quiz` button display only if at least one card has been added to the deck.
Fixed a bug in the reducers, where the `key` property was still looked for when an `addCard` action was dispatched, while along the app that property had been replaced with the card title.