feat: Add score mechanism
Added code to implement score mechanism. The user can submit his/her performance by clicking either the `correct` or `incorrect` button after answering the question.
After the last question, the user score is displayed.
Fixed some bugs in the number of cards already viewed, that did not reach the final expected value equal to the number of cards in the deck.