refactor: Route the user to the New Deck View
Refactored the code to route the user to the Individual Deck View for the new deck, after one had been created.
Change made to comply to the rubric for this project. Previously after a user submitted a new deck, the app navigated back to the default View (the decks list).